requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
boto3==1.28.40
psycopg2-binary==2.9.7
pytest==7.4.0
python-dateutil==2.8.2